+++ bindings/nostr-sdk-js/src/client/signer/mod.rs
@@ -87,7 +87,7 @@ impl JsNostrSigner {
     #[wasm_bindgen(js_name = nip04Encrypt)]
     pub async fn nip04_encrypt(&self, public_key: &JsPublicKey, content: String) -> Result<String> {
         self.inner
-            .nip04_encrypt(**public_key, content)
+            .nip04_encrypt(public_key.deref(), content)
             .await
             .map_err(into_err)
     }
@@ -99,7 +99,7 @@ impl JsNostrSigner {
         encrypted_content: String,
     ) -> Result<String> {
         self.inner
-            .nip04_decrypt(**public_key, encrypted_content)
+            .nip04_decrypt(public_key.deref(), encrypted_content)
             .await
             .map_err(into_err)
     }
@@ -107,7 +107,7 @@ impl JsNostrSigner {
     #[wasm_bindgen(js_name = nip44Encrypt)]
     pub async fn nip44_encrypt(&self, public_key: &JsPublicKey, content: String) -> Result<String> {
         self.inner
-            .nip44_encrypt(**public_key, content)
+            .nip44_encrypt(public_key.deref(), content)
             .await
             .map_err(into_err)
     }
@@ -115,7 +115,7 @@ impl JsNostrSigner {
     #[wasm_bindgen(js_name = nip44Decrypt)]
     pub async fn nip44_decrypt(&self, public_key: &JsPublicKey, content: String) -> Result<String> {
         self.inner
-            .nip44_decrypt(**public_key, content)
+            .nip44_decrypt(public_key.deref(), content)
             .await
             .map_err(into_err)
     }